+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2021-01-21
+## Added
+- `Cache.clear()`
+- Cache option for `clearAfter` that calculates `ttl` from current datetime.
+
+
 ## [1.3.2] - 2020-10-29
 ## Fixed
 - handle empty array for `batchWrite`, `batchGet`, `batchWriteAll` and `batchGetAll`